requirements.txt
psutil
dpkt
scapy
numpy
tensorflow[and-cuda]
keras
scikit-learn
IPython
tqdm
pyyaml
xgboost
catboost
lightgbm
cuml
